sm8550-common: Import android.hardware.power-service.lineage-libperfmgr to device
This needs to be done because we need to remove the vintf entry. We cannot compile this from hardware/lineage/interfaces because if we do it causes a duplicate vintf error with vendor/qcom/opensource/power. This is because we cannot stop qcom/opensource/power from being built as out board is defined as QCOM.... So instead we import power-libperfmgr to device and block qcom power from starting with the previous commit ;)

Change-Id: I2949c65f7b40db9b832f919122261f5ac2b6b8fc